Типизирован слайс tabs
